==> src/miscR/bcmca/1_bcmca_filenames.Rmd <==
---
title: 'BCMCA datasets: import, raster, display'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    toc: true
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: ../src/templates/ohibc_hdr.html
  pdf_document:
    toc: true
---

``` {r setup, echo = FALSE, message = FALSE, warning = FALSE}
library(sp)        # the classes and methods that make up spatial ops in R
library(rgdal)
library(raster)    # call explicitly e.g. raster::extract()
library(DT)

dir_git <- '~/github/ohibc'
dir_rgn <- file.path(dir_git, 'prep/regions')  ### github: general buffer region shapefiles

source(file.path(dir_git, 'src/R/common.R'))  ### an OHIBC specific version of common.R
# source(file.path(dir_git, 'src/R/poly_plot_scores.R')) ### score plotting script

### goal specific folders and info
dir_git_bcmca <- file.path(dir_git, 'bcmca')
dir_anx_bcmca <- file.path(dir_neptune_data, 'git-annex/bcprep/data/bcmca')

### provenance tracking
source('~/github/ohibc/src/R/prov.R')      
  ### Provenance tracking functions: must source at start to initialize prov_track
dir_prov <- file.path(dir_git_bcmca, 'prov') 
  ### set a provenance folder for this script
this_script_file <- file.path(dir_git_bcmca, 'data_prep_spp.Rmd') 
  ### can't recognize the current script name on its own :(
prov_run_tag <- 'standard run'

### set up proj4string options: BC Albers and WGS84
p4s_opts <- c('EPSG:3005 NAD83/BC Albers' = '+proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0',
              'EPSG:4326 WGS 84'          = '+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0')
p4s_bcalb <- p4s_opts[1]
```

# Create list of all BCMCA shapefiles

1. from main folder location on git-annex, identify the major group folders.
    - `dir_neptune_data, 'git-annex/bcprep/data/bcmca'`
    - folders named `bcmca_eco` or `bcmca_hu` for human or ecological data
    - folders then named `_set_xxx_complete` for sets of data
        - only `bcmca_eco_ecologicalfeaturecount_data` is different
    - in sets, subfolders:
        - look for `FeatureData_xxx` where xxx is the group.
        - if no `FeatureData_xxx`, look for `MarxanData_xxx`.
    - in subfolders, list all separate files (look for `.shp` extensions)
    - looks like files have a `BCMCA_ECO` or `BCMCA_HU` prefix
    - looks like case is inconsistent from group to group.
    
``` {r identify all bcmca shapefiles, echo = FALSE, message = FALSE, warning = FALSE}
bcmca_groups <- list.files(dir_anx_bcmca)
bcmca_dirs <- data.frame('dir' = list.dirs(dir_anx_bcmca))

shps_df <- data.frame()
for(dir in bcmca_dirs$dir) {  # i = 4
  if(!is.na(dir)) {
    shps <- list.files(dir)
    shps <- shps[str_detect(shps, '.shp$')]
    if(length(shps) == 0) shps <- NA
    shps <- data.frame('dir' = dir, 'shp' = shps)
  } else {
    shps <- data.frame('dir' = dir, 'shp' = NA)
  }
  shps_df <- bind_rows(shps_df, shps)
}
```

``` {r identify data types and duplicates, echo = FALSE, message = FALSE, warning = FALSE}
### identify the data type (Marxan, feature, pt_feature, poly_feature) 
### from the base name.
shps_df <- shps_df %>%
  filter(!is.na(shp)) %>%
  mutate(data_type = NA, # default: unknown data type
         data_type = ifelse(str_detect(tolower(shp), '_data'),    'feature',      data_type),
         data_type = ifelse(str_detect(tolower(shp), '_marxan'),  'marxan',       data_type),
         data_type = ifelse(str_detect(tolower(shp), '_ptdata'),  'pt_feature',   data_type),
         data_type = ifelse(str_detect(tolower(shp), '_plydata'), 'poly_feature', data_type))

### set up a column of lower case base names with data types removed, to
### compare the basic information for possible duplicates
shps_df <- shps_df %>%
  mutate(base_shp = tolower(shp),
         base_shp = str_replace(base_shp, '_marxan',  ''),
         base_shp = str_replace(base_shp, '_data',    ''),
         base_shp = str_replace(base_shp, '_ptdata',  ''),
         base_shp = str_replace(base_shp, '_plydata', ''))

shps_df$dupes <- duplicated(shps_df$base_shp) | duplicated(shps_df$base_shp, fromLast = TRUE)


### set up columns for domain (ecological or human) and category (e.g. commercialfish etc)
base_split <- str_split(shps_df$base_shp, '_')

get_shp_info <- function(x, y) {
  ifelse(x[1] == 'bcmca', x[y], NA)
}
get_layer <- function(x) {
  ifelse(x[1] == 'bcmca',
         ifelse(length(x) > 3, 
                str_replace(paste(x[4:length(x)], collapse = '_'), '.shp', ''),
                str_replace(x[3], '.shp', '')),
         str_replace(paste(x, collapse = '_'), '.shp', ''))
}

shps_df$domain   <- sapply(base_split, get_shp_info, 2)
shps_df$category <- sapply(base_split, get_shp_info, 3)
shps_df$layer    <- sapply(base_split, get_layer)
  
shps_df <- shps_df %>%
  select(-base_shp)

DT::datatable(shps_df,
              caption  = 'BCMCA shapefiles and locations',
              rownames = FALSE,
              class    = 'stripe hover compact',
              filter   = 'bottom',
              options  = list(dom = 'tpf'))

# write.csv(shps_df, file.path(dir_git_bcmca, 'bcmca_shps_all.csv'), row.names = FALSE)
```


# Read .shps, get info
Get extents, p4s (if not BC Albers), layers available in attribute table; write to .csv.  Set to eval = FALSE to avoid a lengthy process.

``` {r read shps, eval = FALSE}

shps_df <- shps_df %>%
  mutate(extent    = NA,
         data_cols = NA,
         p4s = NA)

for (i in 1:nrow(shps_df)) { # i = 30
  if(is.na(shps_df$shp[i])) {
    message(sprintf('No shapefile available for layer %s.', shps_df$layer[i]))
#   } else if(is.na(shps_df$data_type[i])) {
#     message(sprintf('Layer %s is an ambiguous data layer... figure out later!', shps_df$layer[i]))
#   } else if(str_detect(shps_df$data_type[i], 'pt')) {
#     message(sprintf('Layer %s is a points layer... figure out later!', shps_df$layer[i]))
  } else if(shps_df$layer[i] %in% c('tankerexclusionzone', 'stellersealionhaulouts', 'stellersealionrookeries')) {
    message(sprintf('Layer %s is a problem layer... figure out later!', shps_df$layer[i]))
  } else {
    shp_dir <- shps_df$dir[i]
    shp_lyr <- str_replace(shps_df$shp[i], '.shp', '') 
    lyr_poly <- readOGR(dsn = shp_dir, layer = shp_lyr)
    lyr_p4s  <- proj4string(bcmca_poly)
    
    if(lyr_p4s != p4s_bcalb) {
      message(sprintf('Mismatched projections?\n  Layer p4s: %s\n  BC Albers p4s: %s', shps_df$layer[i], bcmca_p4s, p4s_bcalb))
      shps_df$p4s[i] = lyr_p4s
    } else shps_df$p4s[i] = 'BC Albers'
    
    lyr_ext <- c('xmin' = lyr_poly@bbox[1, 1], 'xmax' = lyr_poly@bbox[1, 2], 'ymin' = lyr_poly@bbox[2, 1], 'ymax' = lyr_poly@bbox[2, 2])
    lyr_data <- names(lyr_poly@data)
    
    shps_df$extent[i]    <- paste(lyr_ext, collapse = ', ')
    shps_df$data_cols[i] <- paste(lyr_data, collapse = ', ')
    
  }
}
shps_df <- shps_df %>%
  separate(extent, c('xmin', 'xmax', 'ymin', 'ymax'), sep = ', ') %>%
  mutate(xmin = as.numeric(xmin),
         xmax = as.numeric(xmax),
         ymin = as.numeric(ymin),
         ymax = as.numeric(ymax))

write.csv(shps_df, file.path(dir_git_bcmca, 'bcmca_shps_all.csv'), row.names = FALSE)

```


``` {r child = file.path(dir_git, 'src/templates/ohibc_prov_ftr.Rmd')}
```
