==> src/miscR/bcmca/2_bcmca_shp_fixes.Rmd <==
---
title: 'BCMCA datasets: import, raster, display'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    toc: true
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: ../src/templates/ohibc_hdr.html
  pdf_document:
    toc: true
---

``` {r setup, echo = FALSE, message = FALSE, warning = FALSE}
library(sp)        # the classes and methods that make up spatial ops in R
library(rgdal)
library(raster)    # call explicitly e.g. raster::extract()
library(DT)

dir_git <- '~/github/ohibc'
dir_rgn <- file.path(dir_git, 'prep/regions')  ### github: general buffer region shapefiles

source(file.path(dir_git, 'src/R/common.R'))  ### an OHIBC specific version of common.R
# source(file.path(dir_git, 'src/R/poly_plot_scores.R')) ### score plotting script

### goal specific folders and info
dir_git_bcmca <- file.path(dir_git, 'bcmca')
dir_anx_bcmca <- file.path(dir_neptune_data, 'git-annex/bcprep/data/bcmca')

### provenance tracking
source('~/github/ohibc/src/R/prov.R')      
  ### Provenance tracking functions: must source at start to initialize prov_track
dir_prov <- file.path(dir_git_bcmca, 'prov') 
  ### set a provenance folder for this script
this_script_file <- file.path(dir_git_bcmca, 'data_prep_spp.Rmd') 
  ### can't recognize the current script name on its own :(
prov_run_tag <- 'standard run'

### set up proj4string options: BC Albers and WGS84
p4s_opts <- c('EPSG:3005 NAD83/BC Albers' = '+proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0',
              'EPSG:4326 WGS 84'          = '+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0')
p4s_bcalb <- p4s_opts[1]
```


# Set up list of layers to be fixed

The identified layers will be "fixed" by reading them in, and dividing the ID field into separate information.

* `benthicclasses` will be divided to depth, substrate, and formation according to the rule for field `ID`:
    * 1000's = depth
        * 1000 <- 0 - 20 m
        * 2000 <- 20 - 50 m
        * 3000 <- 50 - 200 m
        * 4000 <- 200 m +
    * 10's = substrate
        * 10 <- mud
        * 20 <- sand
        * 30 <- hard
        * 90 <- unknown
    * 1's = BPI
        * 1 <- ridge
        * 2 <- depression
        * 3 <- flat
        * 4 <- slope
        
* `coastalclasses` will be determined according to the rule from field `ID`:
    * Initial digit (2) is unused?; delete
    * Next two digits (0 - 34) denote the coastal class according to http://alaskafisheries.noaa.gov/mapping/DataDictionary/Common/ShoreType.htm
    * final one or two digits represents geographic region; delete

* `exposure` will be determined according to the rule from field `ID`:
    * initial two digits (24) are unused?; delete
    * next single digit (1 - 6) denotes exposure
    * final one or two digits represents geographic region; delete


``` {r set up list of raster layers, echo = FALSE, message = FALSE, warning = FALSE}

shps_df <- read.csv(file.path(dir_git_bcmca, 'bcmca_shps_all.csv'), stringsAsFactors = FALSE)


fix_list <- c('benthicclasses', 'coastalclasses', 'exposure')
  
shps_df <- shps_df %>%
  filter(layer %in% fix_list) %>%
  filter(p4s == 'BC Albers')

DT::datatable(shps_df,
              caption  = 'BCMCA shapefiles and locations, files to fix',
              rownames = FALSE,
              class    = 'stripe hover compact',
              options  = list(dom = 'tp'))

```

``` {r deconstruct benthic classes}

### read shp using readOGR
benth <- shps_df %>%
  filter(str_detect(layer, 'benthic') & str_detect(data_type, 'feature'))
### extract ID field into dataframe, and convert ID into character
benth_poly <- readOGR(dsn = benth$dir, layer = str_replace(benth$shp, '.shp', ''), stringsAsFactors = FALSE)
benth_df <- benth_poly@data %>%
  mutate(depth     = as.integer(substr(as.character(Marxan_ID), 1, 1)),
         substrate = as.integer(substr(as.character(Marxan_ID), 3, 3)),
         bpi       = as.integer(substr(as.character(Marxan_ID), 4, 4)))

codes_df <- data.frame('depth' = c(1:4), 'depth_str' = c('0 - 20 m', '20 - 50 m', '50 - 200 m', '200 m +'),
                       'substrate' = c(1, 2, 3, 9), 'subst_str' = c('mud', 'sand', 'hard', 'unknown'),
                       'bpi' = c(1:4), 'bpi_str' = c('ridge', 'depression', 'flat', 'slope'))

benth_df <- benth_df %>%
  left_join(codes_df %>% select(depth, depth_str), by = 'depth') %>%
  left_join(codes_df %>% select(substrate, subst_str), by = 'substrate') %>%
  left_join(codes_df %>% select(bpi, bpi_str), by = 'bpi')

benth_poly@data <- benth_poly@data %>%
  left_join(benth_df, by = 'Marxan_ID')

writeOGR(benth_poly, dsn = file.path(dir_anx_bcmca), layer = str_replace(benth$shp, '.shp', '_mod'), 
         driver = 'ESRI Shapefile',
         overwrite_layer = TRUE)

```

``` {r deconstruct coastal classes}

### read shp using readOGR
coast <- shps_df %>%
  filter(str_detect(layer, 'coastal'))
### extract ID field into dataframe, and convert ID into character
coast_poly <- readOGR(dsn = coast$dir, layer = str_replace(coast$shp, '.shp', ''), stringsAsFactors = FALSE)
coast_df <- coast_poly@data %>%
  mutate(class_id = as.integer(substr(as.character(ID), 2, 3)))

codes_df <- read.csv(file.path(dir_git_bcmca, 'coastal_classes.csv'), stringsAsFactors = FALSE)

coast_df <- coast_df %>%
  left_join(codes_df, by = 'class_id')

coast_poly@data <- coast_poly@data %>%
  left_join(coast_df, by = c('UNIT_ID', 'ID', 'Shape_Leng'))

### problems using writeOGR for some reason...
writeOGR(coast_poly, dsn = file.path(dir_git_bcmca), layer = str_replace(coast$shp, '.shp', '_mod'), 
         driver = 'ESRI Shapefile',
         overwrite_layer = TRUE)
maptools::writeLinesShape(coast_poly, fn = file.path(dir_anx_bcmca, 'modded_lyrs', str_replace(coast$shp, '.shp', '_mod.shp')))

```

``` {r deconstruct exposure classes}

### read shp using readOGR
expos <- shps_df %>%
  filter(str_detect(layer, 'exposure'))
### extract ID field into dataframe, and convert ID into character
expos_poly <- readOGR(dsn = expos$dir, layer = str_replace(expos$shp, '.shp', ''), stringsAsFactors = FALSE)
expos_df <- expos_poly@data %>%
  mutate(expos_id = as.integer(substr(as.character(ID), 3, 3)))

codes_df <- data.frame('expos_id' = c(1:6), 
                       'expos_str' = c('very protected', 'protected', 'semi-protected', 
                                       'semi-exposed',   'exposed',   'very exposed'))

expos_df <- expos_df %>%
  left_join(codes_df, by = 'expos_id')

expos_poly@data <- expos_poly@data %>%
  left_join(expos_df, by = c('UNIT_ID', 'ID', 'Shape_Leng'))

### problems using writeOGR for some reason...
# writeOGR(expos_poly, dsn = file.path(dir_git_bcmca), layer = str_replace(expos$shp, '.shp', '_mod'), 
#          driver = 'ESRI Shapefile',
#          overwrite_layer = TRUE)
maptools::writeLinesShape(expos_poly, fn = file.path(dir_anx_bcmca, 'modded_lyrs', str_replace(expos$shp, '.shp', '_mod.shp')))

```

``` {r child = file.path(dir_git, 'src/templates/ohibc_prov_ftr.Rmd')}
```
