==> src/miscR/bcmca/3_bcmca_rasterize.Rmd <==
---
title: 'BCMCA datasets: import, raster, display'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    toc: true
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: ../src/templates/ohibc_hdr.html
  pdf_document:
    toc: true
---

``` {r setup, echo = FALSE, message = FALSE, warning = FALSE}
library(sp)        # the classes and methods that make up spatial ops in R
library(rgdal)
library(raster)    # call explicitly e.g. raster::extract()
library(DT)

dir_git <- '~/github/ohibc'
dir_rgn <- file.path(dir_git, 'prep/regions')  ### github: general buffer region shapefiles

source(file.path(dir_git, 'src/R/common.R'))  ### an OHIBC specific version of common.R
# source(file.path(dir_git, 'src/R/poly_plot_scores.R')) ### score plotting script

### goal specific folders and info
dir_git_bcmca <- file.path(dir_git, 'bcmca')
dir_anx_bcmca <- file.path(dir_neptune_data, 'git-annex/bcprep/data/bcmca')

### provenance tracking
source('~/github/ohibc/src/R/prov.R')      
  ### Provenance tracking functions: must source at start to initialize prov_track
dir_prov <- file.path(dir_git_bcmca, 'prov') 
  ### set a provenance folder for this script
this_script_file <- file.path(dir_git_bcmca, 'data_prep_spp.Rmd') 
  ### can't recognize the current script name on its own :(
prov_run_tag <- 'standard run'

### set up proj4string options: BC Albers and WGS84
p4s_opts <- c('EPSG:3005 NAD83/BC Albers' = '+proj=aea +lat_1=50 +lat_2=58.5 +lat_0=45 +lon_0=-126 +x_0=1000000 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0',
              'EPSG:4326 WGS 84'          = '+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0')
p4s_bcalb <- p4s_opts[1]
```

# Set up list of layers to be rasterized

``` {r set up list of raster layers, echo = FALSE, message = FALSE, warning = FALSE}

shps_df <- read.csv(file.path(dir_git_bcmca, 'bcmca_shps_all.csv'), stringsAsFactors = FALSE)

rast_list <- shps_df %>%
  left_join(read.csv(file.path(dir_git_bcmca, 'bcmca_to_raster.csv'), stringsAsFactors = FALSE) %>%
              select(shp, raster, value),
            by = 'shp') %>%
  filter(raster == TRUE) %>%
  filter(p4s == 'BC Albers')

DT::datatable(rast_list,
              caption  = 'BCMCA shapefiles and locations, Marxan duplicates removed',
              rownames = FALSE,
              class    = 'stripe hover compact',
              filter   = 'bottom',
              options  = list(dom = 'tpf'))

```

``` {r set up raster extents}

ext_df <- rast_list %>%
  filter(p4s == 'BC Albers') %>%
  filter(layer != 'halibut') %>% ### halibut layer extends up the entire coast; ignore for extents
  mutate(xmin = ifelse(!is.finite(xmin), NA, xmin),
         xmax = ifelse(!is.finite(xmax), NA, xmax),
         ymin = ifelse(!is.finite(ymin), NA, ymin),
         ymax = ifelse(!is.finite(ymax), NA, ymax))

### Marxan reporting grid in BCMCA dataset is offset from perfect 2 km x 2 km grid by: 
x_offset <- 60
y_offset <- 650

reso <- 2000
full_ext <- c('xmin' = round(min(ext_df$xmin, na.rm = TRUE) - 5000, -4) + x_offset, 
              'xmax' = round(max(ext_df$xmax, na.rm = TRUE) + 5000, -4) + x_offset, 
              'ymin' = round(min(ext_df$ymin, na.rm = TRUE) - 5000, -4) + y_offset,
              'ymax' = round(max(ext_df$ymax, na.rm = TRUE) + 5000, -4) + y_offset)

rast_base <- raster::raster(raster::extent(full_ext), resolution = 2000, crs = p4s_bcalb)

```

``` {r rasterize layers}

for(i in 1:nrow(rast_list)) { # i = 1
  tmp <- rast_list[i, ]
  val_list <- unlist(str_split(tmp$value, ','))
  for(val in val_list) {
    lyr_name <- tolower(paste(tmp$category, tmp$layer, val, sep = '_'))

    message(sprintf('Preparing layer %s...', lyr_name))
  
    ### read the layer
    ### use value_id = 1 for all of 'em for now to assign poly@data[ , 1] as values
    ### rasterize the layer to the rast_base
    ### save the raster to git_anx
  
    tmp_data <- foreign::read.dbf(file.path(tmp$dir, str_replace(tmp$shp, '.shp', '.dbf')))
    print(head(tmp_data))
    index_name <- val
    # tmp_filename <- file.path(dir_anx_bcmca, sprintf('rasters/%s.tif', lyr_name))
    tmp_filename <- file.path(dir_git_bcmca, sprintf('rasters/%s.tif', lyr_name))
    suppressWarnings({
      tmp_file <- raster::writeRaster(rast_base, tmp_filename, overwrite = TRUE)
    })
      ### because gdal_rasterize can't create new files... ugh
    tmp_rast <- gdalUtils::gdal_rasterize(src_datasource = file.path(tmp$dir, tmp$shp), 
                                          dst_filename = path.expand(tmp_filename), 
                                          a = index_name, 
                                          a_nodata = NA,
                                          te = c(full_ext[1], full_ext[3], full_ext[2], full_ext[4]),
                                          tr = c(reso, reso), # extents and resolution for x and y
                                  output_Raster = TRUE, # return output as a RasterBrick? 
                                  ignore.full_scan = TRUE,
                                  verbose = FALSE)
    plot(tmp_rast)
  }
}

```

``` {r child = file.path(dir_git, 'src/templates/ohibc_prov_ftr.Rmd')}
```
